pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fielddata-weather-alerts"
version = "0.1.0"
description = "Climate alerts challenge for FieldData"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.14.1",
  "APScheduler==3.10.4",
  "asyncpg==0.30.0",
  "fastapi==0.115.12",
  "pydantic-settings==2.8.1",
  "python-dateutil==2.9.0.post0",
  "sqlalchemy==2.0.40",
  "uvicorn[standard]==0.34.0",
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan==2.1.0",
  "httpx==0.28.1",
  "pytest==8.3.5",
  "pytest-asyncio==0.26.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
